requirements.txt

streamlit
torch
pandas
opencv-python-headless
Pillow
